feat: Enhance frontend UI/UX and mobile responsiveness
This commit introduces a series of frontend improvements focused on creating a more polished, consistent, and responsive user experience.

Key changes:

- **UI/UX Audit and Refinement:**
  - Audited core UI components, including buttons, inputs, and textareas.
  - Removed inconsistent global hover effects on buttons.
  - Standardized focus-visible rings across interactive elements for better accessibility and visual consistency.

- **Enhanced Chat Interface:**
  - Integrated `framer-motion` to add a subtle fade-in and slide-up animation for new chat messages, making the interface more dynamic and engaging.

- **Mobile Responsiveness:**
  - Reworked the main sidebar to be fully responsive.
  - On mobile devices, the sidebar now appears as a full-width, dismissible overlay.
  - Added a hamburger menu icon for toggling the sidebar on small screens, improving the mobile user experience.
