+++ lib/chat-handlers/hosted-chat-handler.ts
@@ -1,0 +1,68 @@
+import {
+  adaptMessagesForGoogleGemini,
+  buildFinalMessages
+} from "@/lib/build-prompt"
+import { Tables } from "@/supabase/types"
+import { ChatPayload, LLM, MessageImage } from "@/types"
+import { fetchChatResponse, processResponse } from "."
+
+export const handleHostedChat = async (
+  payload: ChatPayload,
+  profile: Tables<"profiles">,
+  modelData: LLM,
+  isRegeneration: boolean,
+  newAbortController: AbortController,
+  newMessageImages: MessageImage[],
+  chatImages: MessageImage[],
+  setIsGenerating: React.Dispatch<React.SetStateAction<boolean>>,
+  setFirstTokenReceived: React.Dispatch<React.SetStateAction<boolean>>,
+  setChatMessages: React.Dispatch<React.SetStateAction<any[]>>,
+  setToolInUse: React.Dispatch<React.SetStateAction<string>>
+) => {
+  const provider =
+    modelData.provider === "openai" && profile.use_azure_openai
+      ? "azure"
+      : modelData.provider
+
+  let draftMessages = await buildFinalMessages(payload, profile, chatImages)
+
+  let formattedMessages: any[] = []
+  if (provider === "google") {
+    formattedMessages = await adaptMessagesForGoogleGemini(
+      payload,
+      draftMessages
+    )
+  } else {
+    formattedMessages = draftMessages
+  }
+
+  const apiEndpoint =
+    provider === "custom" ? "/api/chat/custom" : `/api/chat/${provider}`
+
+  const requestBody = {
+    chatSettings: payload.chatSettings,
+    messages: formattedMessages,
+    customModelId: provider === "custom" ? modelData.hostedId : ""
+  }
+
+  const response = await fetchChatResponse(
+    apiEndpoint,
+    requestBody,
+    true,
+    newAbortController,
+    setIsGenerating,
+    setChatMessages
+  )
+
+  return await processResponse(
+    response,
+    isRegeneration
+      ? payload.chatMessages[payload.chatMessages.length - 1]
+      : payload.chatMessages[payload.chatMessages.length - 1],
+    true,
+    newAbortController,
+    setFirstTokenReceived,
+    setChatMessages,
+    setToolInUse
+  )
+}
