+++ lib/chat-helpers.ts
@@ -1,55 +1,11 @@
-// Only used in use-chat-handler.tsx to keep it clean
-
 import { createChatFiles } from "@/db/chat-files"
 import { createChat } from "@/db/chats"
 import { createMessageFileItems } from "@/db/message-file-items"
 import { createMessages, updateMessage } from "@/db/messages"
 import { uploadMessageImage } from "@/db/storage/message-images"
-import {
-  buildFinalMessages,
-  adaptMessagesForGoogleGemini
-} from "@/lib/build-prompt"
-import { consumeReadableStream } from "@/lib/consume-stream"
 import { Tables, TablesInsert } from "@/supabase/types"
-import {
-  ChatFile,
-  ChatMessage,
-  ChatPayload,
-  ChatSettings,
-  LLM,
-  MessageImage
-} from "@/types"
-import React from "react"
-import { toast } from "sonner"
+import { ChatFile, ChatMessage, ChatSettings, LLM, MessageImage } from "@/types"
 import { v4 as uuidv4 } from "uuid"
-
-export const validateChatSettings = (
-  chatSettings: ChatSettings | null,
-  modelData: LLM | undefined,
-  profile: Tables<"profiles"> | null,
-  selectedWorkspace: Tables<"workspaces"> | null,
-  messageContent: string
-) => {
-  if (!chatSettings) {
-    throw new Error("Chat settings not found")
-  }
-
-  if (!modelData) {
-    throw new Error("Model not found")
-  }
-
-  if (!profile) {
-    throw new Error("Profile not found")
-  }
-
-  if (!selectedWorkspace) {
-    throw new Error("Workspace not found")
-  }
-
-  if (!messageContent) {
-    throw new Error("Message content not found")
-  }
-}
 
 export const handleRetrieval = async (
   userInput: string,
@@ -141,212 +97,6 @@ export const createTempMessages = (
   return {
     tempUserChatMessage,
     tempAssistantChatMessage
-  }
-}
-
-export const handleLocalChat = async (
-  payload: ChatPayload,
-  profile: Tables<"profiles">,
-  chatSettings: ChatSettings,
-  tempAssistantMessage: ChatMessage,
-  isRegeneration: boolean,
-  newAbortController: AbortController,
-  setIsGenerating: React.Dispatch<React.SetStateAction<boolean>>,
-  setFirstTokenReceived: React.Dispatch<React.SetStateAction<boolean>>,
-  setChatMessages: React.Dispatch<React.SetStateAction<ChatMessage[]>>,
-  setToolInUse: React.Dispatch<React.SetStateAction<string>>
-) => {
-  const formattedMessages = await buildFinalMessages(payload, profile, [])
-
-  // Ollama API: https://github.com/jmorganca/ollama/blob/main/docs/api.md
-  const response = await fetchChatResponse(
-    process.env.NEXT_PUBLIC_OLLAMA_URL + "/api/chat",
-    {
-      model: chatSettings.model,
-      messages: formattedMessages,
-      options: {
-        temperature: payload.chatSettings.temperature
-      }
-    },
-    false,
-    newAbortController,
-    setIsGenerating,
-    setChatMessages
-  )
-
-  return await processResponse(
-    response,
-    isRegeneration
-      ? payload.chatMessages[payload.chatMessages.length - 1]
-      : tempAssistantMessage,
-    false,
-    newAbortController,
-    setFirstTokenReceived,
-    setChatMessages,
-    setToolInUse
-  )
-}
-
-export const handleHostedChat = async (
-  payload: ChatPayload,
-  profile: Tables<"profiles">,
-  modelData: LLM,
-  tempAssistantChatMessage: ChatMessage,
-  isRegeneration: boolean,
-  newAbortController: AbortController,
-  newMessageImages: MessageImage[],
-  chatImages: MessageImage[],
-  setIsGenerating: React.Dispatch<React.SetStateAction<boolean>>,
-  setFirstTokenReceived: React.Dispatch<React.SetStateAction<boolean>>,
-  setChatMessages: React.Dispatch<React.SetStateAction<ChatMessage[]>>,
-  setToolInUse: React.Dispatch<React.SetStateAction<string>>
-) => {
-  const provider =
-    modelData.provider === "openai" && profile.use_azure_openai
-      ? "azure"
-      : modelData.provider
-
-  let draftMessages = await buildFinalMessages(payload, profile, chatImages)
-
-  let formattedMessages : any[] = []
-  if (provider === "google") {
-    formattedMessages = await adaptMessagesForGoogleGemini(payload, draftMessages)
-  } else {
-    formattedMessages = draftMessages
-  }
-
-  const apiEndpoint =
-    provider === "custom" ? "/api/chat/custom" : `/api/chat/${provider}`
-
-  const requestBody = {
-    chatSettings: payload.chatSettings,
-    messages: formattedMessages,
-    customModelId: provider === "custom" ? modelData.hostedId : ""
-  }
-
-  const response = await fetchChatResponse(
-    apiEndpoint,
-    requestBody,
-    true,
-    newAbortController,
-    setIsGenerating,
-    setChatMessages
-  )
-
-  return await processResponse(
-    response,
-    isRegeneration
-      ? payload.chatMessages[payload.chatMessages.length - 1]
-      : tempAssistantChatMessage,
-    true,
-    newAbortController,
-    setFirstTokenReceived,
-    setChatMessages,
-    setToolInUse
-  )
-}
-
-export const fetchChatResponse = async (
-  url: string,
-  body: object,
-  isHosted: boolean,
-  controller: AbortController,
-  setIsGenerating: React.Dispatch<React.SetStateAction<boolean>>,
-  setChatMessages: React.Dispatch<React.SetStateAction<ChatMessage[]>>
-) => {
-  try {
-    const response = await fetch(url, {
-      method: "POST",
-      body: JSON.stringify(body),
-      signal: controller.signal
-    })
-
-    if (!response.ok) {
-      if (response.status === 404 && !isHosted) {
-        toast.error(
-          "Model not found. Make sure you have it downloaded via Ollama."
-        )
-      }
-
-      const errorData = await response.json()
-
-      toast.error(errorData.message)
-
-      setIsGenerating(false)
-      setChatMessages(prevMessages => prevMessages.slice(0, -2))
-    }
-
-    return response
-  } catch (error: any) {
-    toast.error(error.message)
-    setIsGenerating(false)
-    setChatMessages(prevMessages => prevMessages.slice(0, -2))
-    throw error
-  }
-}
-
-export const processResponse = async (
-  response: Response,
-  lastChatMessage: ChatMessage,
-  isHosted: boolean,
-  controller: AbortController,
-  setFirstTokenReceived: React.Dispatch<React.SetStateAction<boolean>>,
-  setChatMessages: React.Dispatch<React.SetStateAction<ChatMessage[]>>,
-  setToolInUse: React.Dispatch<React.SetStateAction<string>>
-) => {
-  let fullText = ""
-  let contentToAdd = ""
-
-  if (response.body) {
-    await consumeReadableStream(
-      response.body,
-      chunk => {
-        setFirstTokenReceived(true)
-        setToolInUse("none")
-
-        try {
-          contentToAdd = isHosted
-            ? chunk
-            : // Ollama's streaming endpoint returns new-line separated JSON
-              // objects. A chunk may have more than one of these objects, so we
-              // need to split the chunk by new-lines and handle each one
-              // separately.
-              chunk
-                .trimEnd()
-                .split("\n")
-                .reduce(
-                  (acc, line) => acc + JSON.parse(line).message.content,
-                  ""
-                )
-          fullText += contentToAdd
-        } catch (error) {
-          console.error("Error parsing JSON:", error)
-        }
-
-        setChatMessages(prev =>
-          prev.map(chatMessage => {
-            if (chatMessage.message.id === lastChatMessage.message.id) {
-              const updatedChatMessage: ChatMessage = {
-                message: {
-                  ...chatMessage.message,
-                  content: fullText
-                },
-                fileItems: chatMessage.fileItems
-              }
-
-              return updatedChatMessage
-            }
-
-            return chatMessage
-          })
-        )
-      },
-      controller.signal
-    )
-
-    return fullText
-  } else {
-    throw new Error("Response body is null")
   }
 }
