+++ lib/chat-helpers/index.ts
@@ -1,0 +1,30 @@
+import { Tables } from "@/supabase/types"
+import { ChatSettings, LLM } from "@/types"
+
+export const validateChatSettings = (
+  chatSettings: ChatSettings | null,
+  modelData: LLM | undefined,
+  profile: Tables<"profiles"> | null,
+  selectedWorkspace: Tables<"workspaces"> | null,
+  messageContent: string
+) => {
+  if (!chatSettings) {
+    throw new Error("Chat settings not found")
+  }
+
+  if (!modelData) {
+    throw new Error("Model not found")
+  }
+
+  if (!profile) {
+    throw new Error("Profile not found")
+  }
+
+  if (!selectedWorkspace) {
+    throw new Error("Workspace not found")
+  }
+
+  if (!messageContent) {
+    throw new Error("Message content not found")
+  }
+}
