+++ package.json
@@ -71,6 +71,7 @@
     "d3-dsv": "^2.0.0",
     "date-fns": "^2.30.0",
     "endent": "^2.1.0",
+    "framer-motion": "^12.23.12",
     "gpt-tokenizer": "^2.1.2",
     "i18next": "^23.7.16",
     "i18next-resources-to-backend": "^1.2.0",
